fix(ci): restore the 60-line ceiling, drop unbaselined new-file entries
Per Codex review on #941:

- studioBuildStatus.test.tsx grew to 109 lines when the build-bar tests
  were appended in cd437ab; extract them into their own
  StudioBuildBar.test.tsx instead of resealing the ceiling, restoring
  the original 60-line file exactly.
- sweep-scope.ts/.test.ts and the new StudioBuildBar.test.tsx are
  post-freeze files; per docs/module-size-debt.md they should stay out
  of the baseline entirely and ride the flat 500-line hard cap, not get
  frozen at their current small size.
