+++ eslint-rules/module-size-baseline.json
@@ -405,8 +405,6 @@
     "apps/api/src/suggestion-sweep.ts": 533,
     "apps/api/src/suggestions.test.ts": 278,
     "apps/api/src/suggestions.ts": 272,
-    "apps/api/src/sweep-scope.test.ts": 33,
-    "apps/api/src/sweep-scope.ts": 18,
     "apps/api/src/symbol-map.test.ts": 77,
     "apps/api/src/symbol-map.ts": 253,
     "apps/api/src/tab-complete-deploy-flag.test.ts": 26,
@@ -769,7 +767,7 @@
     "apps/web/src/StudioBuildBar.tsx": 44,
     "apps/web/src/StudioBuildHistory.test.tsx": 120,
     "apps/web/src/StudioBuildHistory.tsx": 58,
-    "apps/web/src/studioBuildStatus.test.tsx": 109,
+    "apps/web/src/studioBuildStatus.test.tsx": 60,
     "apps/web/src/StudioChatRail.test.tsx": 248,
     "apps/web/src/StudioChatRail.tsx": 306,
     "apps/web/src/studioChatSheet.test.ts": 22,
